Yay, mp3 decoding actually works :)

+++ mp3/mad/huffman.h
@@ -24,40 +24,40 @@
 
 union huffquad {
   struct {
-    unsigned short final  :  1;
-    unsigned short bits   :  3;
-    unsigned short offset : 12;
+    unsigned int final  :  1;
+    unsigned int bits   :  3;
+    unsigned int offset : 12;
   } ptr;
   struct {
-    unsigned short final  :  1;
-    unsigned short hlen   :  3;
-    unsigned short v      :  1;
-    unsigned short w      :  1;
-    unsigned short x      :  1;
-    unsigned short y      :  1;
+    unsigned int final  :  1;
+    unsigned int hlen   :  3;
+    unsigned int v      :  1;
+    unsigned int w      :  1;
+    unsigned int x      :  1;
+    unsigned int y      :  1;
   } value;
-  unsigned short final    :  1;
+  unsigned int final    :  1;
 };
 
 union huffpair {
   struct {
-    unsigned short final  :  1;
-    unsigned short bits   :  3;
-    unsigned short offset : 12;
+    unsigned int final  :  1;
+    unsigned int bits   :  3;
+    unsigned int offset : 12;
   } ptr;
   struct {
-    unsigned short final  :  1;
-    unsigned short hlen   :  3;
-    unsigned short x      :  4;
-    unsigned short y      :  4;
+    unsigned int final  :  1;
+    unsigned int hlen   :  3;
+    unsigned int x      :  4;
+    unsigned int y      :  4;
   } value;
-  unsigned short final    :  1;
+  unsigned int final    :  1;
 };
 
 struct hufftable {
   union huffpair const *table;
-  unsigned short linbits;
-  unsigned short startbits;
+  unsigned int linbits;
+  unsigned int startbits;
 };
 
 extern union huffquad const *const mad_huff_quad_table[2];
